+++ templates/home.html
@@ -1,0 +1,102 @@
+<!DOCTYPE html>
+<html lang="en">
+<title>RPi - Home Heating System</title>
+<meta charset="UTF-8">¨
+<!--<meta name="viewport" content="width=device-width, initial-scale=1">-->
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
+<link rel="stylesheet" href="https://www.w3schools.com/lib/w3-theme-teal.css">  <!--_black original, many color available!-->
+<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto">
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+<style>
+html,body,h1,h2,h3,h4,h5,h6 {font-family: "Roboto", sans-serif;}
+.w3-sidebar {
+  z-index: 3;
+  width: 250px;
+  top: 43px;
+  bottom: 0;
+  height: inherit;
+}
+</style>
+<body>
+
+<!-- Navbar -->
+<div class="w3-top">
+  <div class="w3-bar w3-theme w3-top w3-left-align w3-large">
+    <a class="w3-bar-item w3-button w3-right w3-hide-large w3-hover-white w3-large w3-theme-l1" href="javascript:void(0)" onclick="w3_open()"><i class="fa fa-bars"></i></a>
+    <a href="home" class="w3-bar-item w3-button w3-theme-l1">RPi-HeatingSys</a>
+    <a href="home" class="w3-bar-item w3-button w3-hide-small w3-hover-white">Home</a>
+    <a href="control" class="w3-bar-item w3-button w3-hide-small w3-hover-white">Control</a>
+    <a href="data" class="w3-bar-item w3-button w3-hide-small w3-hover-white">Data</a>
+    <a href="info" class="w3-bar-item w3-button w3-hide-small w3-hover-white">Info</a>
+  </div>
+</div>
+
+<!-- Sidebar -->
+<nav class="w3-sidebar w3-bar-block w3-collapse w3-large w3-theme-l5 w3-animate-left" id="mySidebar">
+  <a href="javascript:void(0)" onclick="w3_close()" class="w3-right w3-xlarge w3-padding-large w3-hover-black w3-hide-large" title="Close Menu">
+    <i class="fa fa-remove"></i>
+  </a>
+  <h4 class="w3-bar-item"><b>Menu</b></h4>
+  <a class="w3-bar-item w3-button w3-hover-black" href="home">Home</a>
+  <a class="w3-bar-item w3-button w3-hover-black" href="control">Control</a>
+  <a class="w3-bar-item w3-button w3-hover-black" href="data">Data</a>
+  <a class="w3-bar-item w3-button w3-hover-black" href="info">Info</a>
+</nav>
+
+<!-- Overlay effect when opening sidebar on small screens -->
+<div class="w3-overlay w3-hide-large" onclick="w3_close()" style="cursor:pointer" title="close side menu" id="myOverlay"></div>
+
+<!-- START MAIN -->
+<!-- shift it to the right by 250 pixels when the sidebar is visible -->
+<div class="w3-main" style="margin-left:250px">
+	
+	<div class="w3-row w3-padding-64">
+	<div class="w3-container">
+	  <h2>Latest Records</h2>
+	  <p>Latest sensor and controller status (current time: {{currentTime}})</p>
+		{{tableHTML_1|safe}}
+	</div>
+	</div>	
+	
+<!-- FOOTER -->	
+  <footer id="myFooter">
+    <div class="w3-container w3-theme-l2 w3-padding-32">
+      <h4>Heating System Control and Web App Design by Benoit Picard <a href="mailto:picard.benoit@gmail.com">picard.benoit@gmail.com</a></h4>
+    </div>
+
+    <div class="w3-container w3-theme-l1">
+      <p>Powered by <a href="https://www.w3schools.com/w3css/default.asp" target="_blank">w3.css</a></p>
+    </div>
+  </footer>
+
+<!-- END MAIN -->
+</div>
+
+<script>
+// Get the Sidebar
+var mySidebar = document.getElementById("mySidebar");
+
+// Get the DIV with overlay effect
+var overlayBg = document.getElementById("myOverlay");
+
+// Toggle between showing and hiding the sidebar, and add overlay effect
+function w3_open() {
+  if (mySidebar.style.display === 'block') {
+    mySidebar.style.display = 'none';
+    overlayBg.style.display = "none";
+  } else {
+    mySidebar.style.display = 'block';
+    overlayBg.style.display = "block";
+  }
+}
+
+// Close the sidebar with the close button
+function w3_close() {
+  mySidebar.style.display = "none";
+  overlayBg.style.display = "none";
+}
+</script>
+
+</body>
+</html>
